lisitng widget and vote, follow controller update for separate action

+++ app/controllers/follows_controller.rb
@@ -1,26 +1,14 @@
 class FollowsController < ApplicationController
 
   before_action :authenticate_user!
-  after_action :verify_authorized, except: [:index]
-  rescue_from Pundit::NotAuthorizedError, with: :user_not_authorized
   before_action :load_followable
 
-  def create
-    if @followable.followed_by?(current_user)
-     unfollow
-    else
-     follow
-    end
-  end
-
   def follow
+    return render_data if @followable.followed_by?(current_user)
     @follow = current_user.followings.new(followable: @followable)
-    authorize @follow.follower
+    authorize @follow
     if @follow.save
-      render json: {
-        follow: @followable.followed_by?(current_user),
-        followers_count: @followable.followers_counter.value
-      }
+      render_data
     else
       respond_to do |format|
         format.json {render json: @follow.errors, status: unprocessable_entity}
@@ -29,13 +17,11 @@ class FollowsController < ApplicationController
   end
 
   def unfollow
+    return render_data if !@followable.followed_by?(current_user)
     @follow = current_user.followings.where(followable: @followable).first
     authorize @follow.follower
     if @follow.destroy
-      render json: {
-        follow: @followable.followed_by?(current_user),
-        followers_count: @followable.followers_counter.value
-      }
+      render_data
     else
       respond_to do |format|
         format.json {render json: @follow.errors, status: unprocessable_entity}
@@ -44,20 +30,33 @@ class FollowsController < ApplicationController
   end
 
   def followers
-    @followers = @followable.followers
+    @followers = @followable.get_followers.paginate(:page => params[:page], :per_page => 9)
     render 'follows/followers'
+  end
+
+  def followings
+    @followings = @followable.get_followings.paginate(:page => params[:page], :per_page => 9)
+    render 'follows/followings'
   end
 
   private
 
+  def render_data
+    render json: Oj.dump({
+      follow: @followable.followed_by?(current_user),
+      followers_count: @followable.followers_counter.value,
+      follower: current_user.card_json
+    })
+  end
+
   def load_followable
-    @follwables = ["Idea", "School", "User"]
-    if @follwables.include? params[:followable_type]
+    @followables = ["Idea", "School", "User"]
+    if @followables.include? params[:followable_type]
       followable_type = params[:followable_type]
-      unless followable_type.safe_constantize.find(params[:followable_id]).blank?
-        @followable = followable_type.safe_constantize.find(params[:followable_id])
-      else
+      if followable_type.safe_constantize.find(params[:followable_id]).blank?
         raise ActiveRecord::RecordNotFound
+      else
+        @followable = followable_type.safe_constantize.find(params[:followable_id])
       end
     else
       respond_to do |format|
